refactor: Move search_term from filters to request structs
Move search_term out of InvoiceFilters and CreditNoteFilter into their
respective request structs (ListInvoicesRequest, ListCustomerInvoicesRequest,
ListCreditNotesRequest) to align with how the Lago API treats search_term
as a top-level parameter rather than a filter.
