+++ lago-client/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "lago-client"
-version = "0.1.19"
+version = "0.1.20"
 edition = "2024"
 authors = ["Lago Team <tech@getlago.com>"]
 description = "Lago API client"
@@ -8,7 +8,7 @@ license = "MIT"
 repository = "https://github.com/getlago/lago-rust-client"
 
 [dependencies]
-lago-types = "0.1.19"
+lago-types = "0.1.20"
 serde = { workspace = true }
 serde_json = { workspace = true }
 tokio = { workspace = true }
