+++ lago-client/examples/invoice.rs
@@ -1,10 +1,7 @@
 use lago_client::LagoClient;
-use lago_types::{
-    filters::invoice::InvoiceFilters,
-    requests::invoice::{
-        DownloadInvoiceRequest, GetInvoiceRequest, ListCustomerInvoicesRequest,
-        ListInvoicesRequest, UpdateInvoiceInput, UpdateInvoiceMetadataInput, UpdateInvoiceRequest,
-    },
+use lago_types::requests::invoice::{
+    DownloadInvoiceRequest, GetInvoiceRequest, ListCustomerInvoicesRequest, ListInvoicesRequest,
+    UpdateInvoiceInput, UpdateInvoiceMetadataInput, UpdateInvoiceRequest,
 };
 
 #[tokio::main]
@@ -17,8 +14,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     println!("Found {} invoices", invoices.invoices.len());
 
     // Example 1b: Search invoices by number, customer name, or email
-    let search_request = ListInvoicesRequest::new()
-        .with_filters(InvoiceFilters::new().with_search_term("LAR-F878-202501".to_string()));
+    let search_request = ListInvoicesRequest::new().with_search_term("LAR-F878-202501".to_string());
     let search_results = client.list_invoices(Some(search_request)).await?;
     println!(
         "Found {} invoices matching search term",
